+++ Source/NLedger/MainApplicationContext.cs
@@ -6,6 +6,8 @@
 // Copyright (c) 2003-2018, John Wiegley.  All rights reserved.
 // See LICENSE.LEDGER file included with the distribution for details and disclaimer.
 // **********************************************************************************
+using NLedger.Abstracts;
+using NLedger.Abstracts.Impl;
 using NLedger.Commodities;
 using NLedger.Formatting;
 using NLedger.Scopus;
@@ -65,9 +67,20 @@ namespace NLedger
         public string TimeZoneId { get; set; }
         public TimeZoneInfo TimeZone { get; set; }
 
+        // Abstract Application Services
+        public IQuoteProvider QuoteProvider => _QuoteProvider.Value;
+        public IProcessManager ProcessManager => _ProcessManager.Value;
+
+        public void SetQuoteProvider(Func<IQuoteProvider> quoteProviderFactory)
+        {
+            _QuoteProvider = new Lazy<IQuoteProvider>(quoteProviderFactory);
+        }
+
         [ThreadStatic]
         private static MainApplicationContext CurrentInstance;
 
         private CommodityPool _CommodityPool;
+        private Lazy<IQuoteProvider> _QuoteProvider = new Lazy<IQuoteProvider>(() => new QuoteProvider());
+        private Lazy<IProcessManager> _ProcessManager = new Lazy<IProcessManager>(() => new ProcessManager());
     }
 }